Set span status to error on exception
Previously we didn't do anything on exception. This meant that the
parent process could have marked the parent span with error code. But
also, if this was ran asynchrounously and there was no error handling,
then this information could have been completely lost.

Here if we know that there's an error then we can immediately add the
error status to the span. This makes it easier to pinpoint where exactly
an error happened.